Use blaze-html to render the templates
This introduces a cabal configure warning:

> Warning: This package indirectly depends on multiple versions of the same
> package. This is highly likely to cause a compile failure.
> package xmlhtml-0.1.0.1 requires blaze-html-0.3.2.1
> package pastie-0.1 requires blaze-html-0.4.0.0

The dependency chain being Pastie -> Snap -> Heist -> XmlHtml.

All use of Heist has been removed from Pastie though. The dependency on Snap
is for 1) the development server and 2) the Snap extension framework.

XmlHtml depends on Blaze Html in order to provide a module implementing a
Blaze Html to XmlHtml renderer. This module is not imported by Heist, so the
version of Blaze Html used by XmlHtml should not end up in Pastie.

If this proves to be a problem it should be possible to fork Snap and remove
the Heist extension from it.