Aborted work on String -> ByteString refactor
I wanted to change the type as I was using unpack in a lot of places which
felt unclean. Also, some of the function that I would have expected to be more
costly, such as highlightAs are operating on Strings.

Whilst it worked, it has been aborted due to:

 - still having to use unpack all over the place
 - the perceived expensive functions still operating on Strings

The expensive functions are operating on Strings as they don't provide an API
to do otherwise. unpack is being called in a lot of places to support APIs
which don't operate on ByteStrings.

One of the APIs which doesn't operate on ByteStrings is Text.Blaze. It does
provide an API for operating on Data.Text (unicode packed strings). This
suggests that any further attempts to replace String with a packed format
should probably be using Data.Text.